Abort recording/parsing conflicts
If a data category is directly recorded after created by parsing rules,
it harms the parsing pipeline; is the category the source of data or
derived from other categories by rules? This case must be avoided.
Also, adding a parsing rule that derives the category directly recorded
before must be avoided.

Signed-off-by: Yunjae Lee <lyj7694@gmail.com>