Add default tag value feature
Two types of the default tags are implemented: implicit and explicit.
Users can set/unset the default value of a tag using `tagit fix`
explicitly. In addition, the default value of the tag is set to the last
value of the tag. The explicit default value has the higher priority
than the implicit one.

Signed-off-by: Yunjae Lee <lyj7694@gmail.com>